+++ providers/darwin/process_darwin_amd64.go
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+// +build darwin,amd64,cgo
+
 package darwin
 
 // #cgo LDFLAGS:-lproc
